+++ src/tables/mod.rs
@@ -20,3 +20,5 @@ pub mod maxp;
 pub mod name;
 pub mod os2;
 pub mod post;
+pub mod vhea;
+pub mod vmtx;